requirements.txt
python-genai
python-dotenv
pandas
datasets
